+++ tests/test_data/allow_wrappers/dangerous-dep/src/lib.rs
@@ -1,0 +1,7 @@
+pub fn panic_if_one(v: u32) -> u32 {
+    if v == 1 {
+        panic!("oops");
+    } else {
+        v
+    }
+}