+++ tests/test_data/allow_wrappers/safe-wrapper/src/lib.rs
@@ -1,0 +1,7 @@
+pub fn safety_first(v: u32) -> u32 {
+    if v != 1 {
+        dangerous_dep::panic_if_one(v)
+    } else {
+        v
+    }
+}